+++ src/Amadeus/Client/RequestCreator/Converter/Fare/PricePNRWithBookingClassConv.php
@@ -1,0 +1,50 @@
+<?php
+/**
+ * amadeus-ws-client
+ *
+ * Copyright 2015 Amadeus Benelux NV
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ * @package Amadeus
+ * @license https://opensource.org/licenses/Apache-2.0 Apache 2.0
+ */
+
+namespace Amadeus\Client\RequestCreator\Converter\Fare;
+
+use Amadeus\Client\RequestCreator\Converter\BaseConverter;
+use Amadeus\Client\RequestOptions\FarePricePnrWithBookingClassOptions;
+use Amadeus\Client\Struct;
+
+/**
+ * Fare_PricePNRWithBookingClass Request converter
+ *
+ * @package Amadeus\Client\RequestCreator\Converter\Fare
+ * @author Dieter Devlieghere <dieter.devlieghere@benelux.amadeus.com>
+ */
+class PricePNRWithBookingClassConv extends BaseConverter
+{
+    /**
+     * @param FarePricePnrWithBookingClassOptions $requestOptions
+     * @param int|string $version
+     * @return Struct\Fare\PricePNRWithBookingClass12|Struct\Fare\PricePNRWithBookingClass13
+     */
+    public function convert($requestOptions, $version)
+    {
+        if ($version < 13) {
+            return new Struct\Fare\PricePNRWithBookingClass12($requestOptions);
+        } else {
+            return new Struct\Fare\PricePNRWithBookingClass13($requestOptions);
+        }
+    }
+}